Build: Fixed compile error for the mbedtls-2.7.9 on Cygwin platform
This patch fixes an issue observed when building Mbed TLS on Cygwin as
part of the TF-M build, related to a failure in the creation of a
symbolic link. The fix forces the build system option -DCMAKE_HOST_UNIX
to ON in the build system settings to avoid this failure.

Change-Id: I92c25fca7220e2b279e1c07b6c3f29f4161546e2
Signed-off-by: Qixiang Xu <qixiang.xu@arm.com>